oops

+++ db/migrate/20121124044754_create_weather_data.rb
@@ -1,12 +1,12 @@
 class CreateWeatherData < ActiveRecord::Migration
   def change
     create_table :weather_data do |t|
-      t.integer :temp
-      t.integer :temp_low
-      t.integer :temp_high
-      t.integer :visibility
-      t.integer :humidity
-      t.integer :pressure
+      t.float :temp
+      t.float :temp_low
+      t.float :temp_high
+      t.float :visibility
+      t.float :humidity
+      t.float :pressure
 
       t.timestamps
     end
